requirements.txt
amqp==5.3.1
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
binaryornot==0.4.4
boto3==1.36.2
botocore==1.36.2
build==1.2.2.post1
celery==5.4.0
celery-progress==0.4
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-allauth==65.3.1
django-anymail==12.0
django-celery-beat==2.7.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-hijack==3.7.1
django-htmx==1.21.0
django-otp==1.5.4
django-storages==1.14.4
django-template-partials==24.4
django-timezone-field==7.1
django-waffle==4.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
fido2==1.2.0
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pegasus-cli==0.7
pip-tools==7.4.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
referencing==0.36.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
zope.interface==7.2
